more structured code with a header and a makefile
